# api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Dynamis Code Apps Template API",
    "version": "1.0.0",
    "description": "Stable v1 REST contract. Additive changes preserve authorization, ordering, errors, and side effects. Breaking changes require a new version or published migration window. Query collections use bounded filters, literal search, deterministic sorting, and opaque query-bound cursors. Errors use RFC 9457 Problem Details. See docs/api.md for the human-readable reference."
  },
  "paths": {
...
    "/api/v1/auth/mfa/status": {
      "get": {
        "operationId": "mfaStatus", "security": [{"sessionCookie": []}],
        "responses": {"200": {"description": "MFA status without secrets", "content": {"application/json": {"schema": {"type": "object"}}}}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/export": {
      "get": {
        "operationId": "exportWorkspace",
        "description": "Returns a bounded, point-in-time workspace export. Authentication secrets and transient records are excluded.",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "responses": {
          "200": {"description": "Workspace export", "headers": {"Content-Disposition": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceExport"}}}},
          "401": {"$ref": "#/components/responses/Problem"},
          "403": {"$ref": "#/components/responses/Problem"},
          "409": {"$ref": "#/components/responses/Problem"},
          "429": {"$ref": "#/components/responses/RateLimited"}
        }
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/members": {
      "get": {
        "operationId": "listMembers", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "responses": {"200": {"description": "Workspace members", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MemberPage"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/members/{userId}": {
      "patch": {
        "operationId": "changeMemberRole", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/UserId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MemberRoleRequest"}}}},
        "responses": {"204": {"description": "Member role changed"}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      },
      "delete": {
        "operationId": "removeMember", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/UserId"}],
        "responses": {"204": {"description": "Member removed"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/ownership": {
      "post": {
        "operationId": "transferOwnership", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OwnershipRequest"}}}},
        "responses": {"204": {"description": "Ownership transferred"}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations": {
      "get": {
        "operationId": "listInvitations", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "responses": {"200": {"description": "Workspace invitations", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InvitationPage"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      },
      "post": {
        "operationId": "createInvitation", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateInvitationRequest"}}}},
        "responses": {"201": {"description": "Invitation created", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InvitationLinkResponse"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations/{invitationId}/resend": {
      "post": {
        "operationId": "resendInvitation", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/InvitationId"}],
        "responses": {"200": {"description": "Invitation resent", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InvitationLinkResponse"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations/{invitationId}": {
      "delete": {
        "operationId": "revokeInvitation", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/InvitationId"}],
        "responses": {"204": {"description": "Invitation revoked"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "operationId": "listSessions", "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Current user's sessions without secrets", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionPage"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/sessions/{sessionId}": {
      "delete": {
        "operationId": "revokeSession", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/SessionId"}],
        "responses": {"204": {"description": "Session revoked"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "session"}
    },
    "parameters": {
      "WorkspaceId": {"name": "workspaceId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
...
      "UserId": {"name": "userId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
      "InvitationId": {"name": "invitationId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
...
      "SessionId": {"name": "sessionId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
...
    },
    "responses": {
      "Problem": {"description": "Problem Details", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
...
      "RateLimited": {"description": "Rate limit exceeded", "headers": {"Retry-After": {"required": true, "schema": {"type": "integer", "minimum": 1}}}, "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
...
      "Item": {"type": "object", "required": ["id", "workspaceId", "createdByUserId", "title", "status", "version", "createdAt", "updatedAt"], "properties": {"id": {"type": "string"}, "workspaceId": {"type": "string"}, "createdByUserId": {"type": ["string", "null"]}, "title": {"type": "string"}, "status": {"type": "string", "enum": ["active", "complete"]}, "version": {"type": "integer", "minimum": 1}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}}},
...
      "ExportWorkspace": {"type": "object", "required": ["id", "name", "createdAt"], "properties": {"id": {"type": "string"}, "name": {"type": "string"}, "locale": {"type": "string", "enum": ["en", "es"]}, "createdAt": {"type": "string", "format": "date-time"}}},
      "ExportMember": {"type": "object", "required": ["userId", "email", "role", "createdAt"], "properties": {"userId": {"type": "string"}, "email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["owner", "admin", "member", "viewer"]}, "createdAt": {"type": "string", "format": "date-time"}}},
      "ExportAuditEvent": {"type": "object", "required": ["id", "eventType", "authMethod", "targetType", "action", "outcome", "metadata", "createdAt"], "properties": {"id": {"type": "string"}, "eventType": {"type": "string"}, "actorUserId": {"type": "string"}, "authMethod": {"type": "string"}, "workspaceId": {"type": "string"}, "targetType": {"type": "string"}, "targetId": {"type": "string"}, "action": {"type": "string"}, "outcome": {"type": "string"}, "requestId": {"type": "string"}, "sourceAddress": {"type": "string"}, "metadata": {}, "createdAt": {"type": "string", "format": "date-time"}}},
      "WorkspaceExport": {"type": "object", "required": ["formatVersion", "exportedAt", "workspace", "members", "items", "auditEvents", "excluded"], "properties": {"formatVersion": {"const": "dynamis-code.workspace/v1"}, "exportedAt": {"type": "string", "format": "date-time"}, "workspace": {"$ref": "#/components/schemas/ExportWorkspace"}, "members": {"type": "array", "items": {"$ref": "#/components/schemas/ExportMember"}}, "items": {"type": "array", "items": {"$ref": "#/components/schemas/Item"}}, "auditEvents": {"type": "array", "items": {"$ref": "#/components/schemas/ExportAuditEvent"}}, "excluded": {"type": "array", "items": {"type": "string"}}}},
...
      "Member": {"type": "object", "required": ["userId", "email", "role", "createdAt"], "properties": {"userId": {"type": "string"}, "email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["owner", "admin", "member", "viewer"]}, "createdAt": {"type": "string", "format": "date-time"}}},
      "MemberPage": {"type": "object", "required": ["members"], "properties": {"members": {"type": "array", "items": {"$ref": "#/components/schemas/Member"}}}},
      "MemberRoleRequest": {"type": "object", "additionalProperties": false, "required": ["role"], "properties": {"role": {"type": "string", "enum": ["admin", "member", "viewer"]}}},
      "OwnershipRequest": {"type": "object", "additionalProperties": false, "required": ["userId"], "properties": {"userId": {"type": "string"}}},
...
      "InvitationPage": {"type": "object", "required": ["invitations"], "properties": {"invitations": {"type": "array", "items": {"$ref": "#/components/schemas/Invitation"}}}},
      "CreateInvitationRequest": {"type": "object", "additionalProperties": false, "required": ["email", "role"], "properties": {"email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["admin", "member", "viewer"]}, "lifetimeSeconds": {"type": "integer", "minimum": 60, "maximum": 2592000}}},
      "InvitationLinkResponse": {"type": "object", "required": ["invitation", "invitationUrl", "emailDelivered"], "properties": {"invitation": {"$ref": "#/components/schemas/Invitation"}, "invitationUrl": {"type": "string", "format": "uri-reference"}, "emailDelivered": {"type": "boolean"}}},
...
      "SessionPage": {"type": "object", "required": ["sessions"], "properties": {"sessions": {"type": "array", "items": {"$ref": "#/components/schemas/Session"}}}},
...
      "Problem": {"type": "object", "required": ["type", "title", "status", "detail", "instance", "code", "requestId"], "properties": {"type": {"type": "string", "format": "uri"}, "title": {"type": "string"}, "status": {"type": "integer"}, "detail": {"type": "string"}, "instance": {"type": "string"}, "code": {"type": "string"}, "requestId": {"type": "string"}}}
    }
  }
}